Feat : Implement a button to make a comprehensive data file UI에 파일 하나로 통합하는 버튼 추가
